+++ src/main/java/ru/practicum/shareit/user/UserDto/UserDto.java
@@ -10,13 +10,13 @@ import java.util.Objects;
 @Data
 @NoArgsConstructor
 public class UserDto {
-    private Integer id;
+    private Long id;
     private String name;
     @NotBlank(message = "Поле email не должно быть пустым")
     @Email(message = "Поле email должно быть заполнено как адрес электронной почты")
     private String email;
 
-    public UserDto(Integer id, String name, String email) {
+    public UserDto(Long id, String name, String email) {
         this.id = id;
         if (name.isBlank()) {
             this.name = email;
